Build libgcc_s DLLs with additional linker flags
Bug: http://b/113171785

This hacked up change passes security-related linker flags for libgcc_s
DLLs.  It should affect only Windows targets in gcc-4.8.3.

There doesn't seem a way to pass separate LDFLAGS for 32- and 64-bit
when running 'configure', so we'd end up without --high-entropy-va for
the 64-bit DLL.  Instead, this hacky change gets the behavior we want
and is unlikely to regress.  To be safe
https://android-review.googlesource.com/c/toolchain/mingw/+/743702 also
verifies that all DLLs built have the correct flags.

Test: toolchain/mingw/build.sh
Change-Id: I9d52e7b8cb449fcaf11c7f771761f11c7d35f99b